V1.0
Feature Complete?

/setloggingchannel channel
Sets specified channel to receive all bot logs.

/setdefaultbulletin channel
Sets specified channel to the default bulletin board channel (pin overflow)

/register channel1 channel2(optional)
Channel1 is the channel you want it to monitor the pins of
Channel2 will redirect the overflow pins somewhere else. If channel2 is not set, they will be sent to the default bulletin board channel.

/lock message_url
Locks a message to the top of the pin list. If multiple messages are locked, it will ensure they are all at the top of the pin list but it may not be kept in order.

/list channel
Lists all the locked pins in a channel.